Move all commands into the root package
