+++ cmd/svcat/binding/bind_cmd.go
@@ -3,9 +3,9 @@ package binding
 import (
 	"fmt"
 
-	"github.com/Azure/service-catalog-cli/pkg/command"
-	"github.com/Azure/service-catalog-cli/pkg/output"
-	"github.com/Azure/service-catalog-cli/pkg/parameters"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/command"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/output"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/parameters"
 	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/spf13/cobra"
 )
@@ -92,6 +92,5 @@ func (c *bindCmd) bind() error {
 	}
 
 	output.WriteBindingDetails(c.Output, binding)
-
 	return nil
 }
