+++ cmd/svcat/binding/unbind_cmd.go
@@ -3,7 +3,7 @@ package binding
 import (
 	"fmt"
 
-	"github.com/Azure/service-catalog-cli/pkg/command"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/command"
 	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/spf13/cobra"
 )
