+++ cmd/svcat/instance/deprovision_cmd.go
@@ -3,7 +3,7 @@ package instance
 import (
 	"fmt"
 
-	"github.com/Azure/service-catalog-cli/pkg/command"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/command"
 	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/spf13/cobra"
 )
