+++ cmd/svcat/main.go
@@ -4,14 +4,14 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/Azure/service-catalog-cli/pkg/binding"
-	"github.com/Azure/service-catalog-cli/pkg/broker"
-	"github.com/Azure/service-catalog-cli/pkg/class"
-	"github.com/Azure/service-catalog-cli/pkg/command"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/binding"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/broker"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/class"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/command"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/instance"
+	"github.com/Azure/service-catalog-cli/cmd/svcat/plan"
 	"github.com/Azure/service-catalog-cli/pkg/environment"
-	"github.com/Azure/service-catalog-cli/pkg/instance"
 	"github.com/Azure/service-catalog-cli/pkg/kube"
-	"github.com/Azure/service-catalog-cli/pkg/plan"
 	"github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset"
 	"github.com/spf13/cobra"
 	"k8s.io/client-go/rest"
