+++ cmd/svcat/output/binding.go
